# api/docs/swagger.yaml
definitions:
  models.Capacity:
    properties:
      cpu:
        type: number
      memory:
        type: integer
    type: object
  models.Catalog:
    properties:
      capacity:
        $ref: '#/definitions/models.Capacity'
      description:
        type: string
      expiry:
        type: integer
      id:
        type: string
      image_thumbnail_reference:
        type: string
      name:
        type: string
      retired:
        type: boolean
      status:
        $ref: '#/definitions/models.CatalogStatus'
      type:
        type: string
      vm:
        $ref: '#/definitions/models.VM'
    type: object
  models.CatalogStatus:
    properties:
      message:
        type: string
      ready:
        type: boolean
    type: object
...
  models.VM:
    properties:
      capacity:
        $ref: '#/definitions/models.Capacity'
      crn:
        type: string
      image:
        type: string
      network:
        type: string
      processor_type:
        type: string
      system_type:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: Go-server for PAC backend apis in gin
  title: Pac-go-server APIs
paths:
...
  /api/v1/catalogs:
    get:
      consumes:
      - application/json
      description: Get all catalogs resource
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all catalogs
      tags:
      - catalogs
    post:
      consumes:
      - application/json
      description: Create catalog resource
      parameters:
      - description: Create catalog
        in: body
        name: catalog
        required: true
        schema:
          $ref: '#/definitions/models.Catalog'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Create catalog
      tags:
      - catalogs
  /api/v1/catalogs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete catalog resource
      parameters:
      - description: catalog name to be deleted
        in: path
        name: name
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete catalog
      tags:
      - catalogs
    get:
      consumes:
      - application/json
      description: Get catalog resource
      parameters:
      - description: catalog name to be fetched
        in: path
        name: name
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get catalog as specified in request
      tags:
      - catalogs
  /api/v1/catalogs/{name}/retire:
    put:
      consumes:
      - application/json
      description: Reire catalog resource
      parameters:
      - description: catalog name to be retired
        in: path
        name: name
        required: true
        type: string
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Reire catalog
      tags:
      - catalogs
...
swagger: "2.0"